requirements.txt
gspread
oauth2client
kafka-python
pandas
python-dotenv
pymongo
numpy
scikit-learn